+++ tests/testthat/test-get_datagrid.R
@@ -180,9 +180,22 @@ test_that("get_datagrid - data", {
   expect_identical(dim(get_datagrid(iris, by = 1:2, length = c(3, 2))), c(6L, 5L))
   expect_identical(dim(get_datagrid(iris, by = 1:2, length = c(NA, 2))), c(70L, 5L))
   expect_identical(dim(get_datagrid(iris, by = "Sepal.Length = c(1, 2)", length = NA)), c(2L, 5L))
-  expect_error(get_datagrid(iris, by = 1:2, length = c(3, 2, 4)))
-  expect_error(get_datagrid(iris, by = 1:2, length = "yes"))
+  expect_error(
+    get_datagrid(iris, by = 1:2, length = c(3, 2, 4)),
+    regex = "The number of"
+  )
+  expect_error(
+    get_datagrid(iris, by = 1:2, length = "yes"),
+    regex = "`length` argument must be a finite",
+    fixed = TRUE
+  )
+  expect_error(
+    get_datagrid(iris, by = 1:2, length = Inf),
+    regex = "`length` argument must be a finite",
+    fixed = TRUE
+  )
   expect_identical(as.numeric(get_datagrid(iris, by = 1:2, range = c("range", "mad"), length = c(2, 3))[4, "Sepal.Width"]), median(iris$Sepal.Width)) # nolint
+  expect_identical(dim(get_datagrid(mtcars, by = "gear", length = NA)), c(3L, 11L))
 
   expect_identical(nrow(get_datagrid(data.frame(
     X = c("A", "A", "B"),
@@ -549,6 +562,7 @@ test_that("get_datagrid - handle integers", {
     y = letters[1:8]
   )
   expect_identical(dim(get_datagrid(d)), c(64L, 2L)) # 8^2
+  expect_identical(dim(get_datagrid(d, length = 5)), c(40L, 2L)) # 5 * 8
   expect_identical(dim(get_datagrid(d, protect_integers = FALSE)), c(80L, 2L)) # 8 * length
   expect_identical(dim(get_datagrid(d, range = "grid")), c(64L, 2L))
   expect_identical(dim(get_datagrid(d, length = 5)), c(40L, 2L))